requirement.txt
aniso8601==8.0.0
click==7.1.2
Flask-RESTful==0.3.8
pytz==2020.1
six==1.15.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.1
gunicorn==21.2.0